+++ src/test/resources/json/ccdResponseWithJointParty.json
@@ -4,6 +4,7 @@
     "callback_response_status": null,
     "callback_response_status_code": null,
     "case_data": {
+      "directionTypeDl": "provideInformation",
       "createdInGapsFrom": "readyToList",
       "jointParty": "Yes",
       "jointPartyName": {
